+++ crates/ecstore/src/set_disk/ops/mod.rs
@@ -21,3 +21,4 @@ pub(crate) mod bucket;
 pub(crate) mod heal;
 pub(crate) mod list;
 pub(crate) mod multipart;
+pub(crate) mod object;
